Move `cmd/internal` to `internal`
This is to help with acceptance test code coverage in #18: `main.go`
needs access to internal packages, and if they're within `cmd` it
doesn't have the access.
