+++ internal/applicationsnapshot/report.go
