+++ internal/applicationsnapshot/report_test.go
