+++ internal/applicationsnapshot/test_snapshot.json
