+++ internal/pipeline/evaluation_target.go
@@ -1,7 +1,7 @@
 package pipeline
 
 import (
-	"github.com/hacbs-contract/ec-cli/cmd/internal/utils"
+	"github.com/hacbs-contract/ec-cli/internal/utils"
 	"github.com/spf13/afero"
 )
