+++ internal/pipeline/evaluator_test.go
@@ -6,7 +6,7 @@ import (
 	"testing"
 
 	"github.com/go-git/go-git/v5"
-	"github.com/hacbs-contract/ec-cli/cmd/internal/utils"
+	"github.com/hacbs-contract/ec-cli/internal/utils"
 	"github.com/open-policy-agent/conftest/runner"
 	"github.com/spf13/afero"
 )
