+++ internal/pipeline/test_data/pipeline_definitions/invalid.txt
