+++ internal/pipeline/test_data/pipeline_definitions/pipeline.json
