+++ internal/pipeline/validate.go
@@ -3,7 +3,7 @@ package pipeline
 import (
 	"context"
 
-	"github.com/hacbs-contract/ec-cli/cmd/internal/policy"
+	"github.com/hacbs-contract/ec-cli/internal/policy"
 	"github.com/open-policy-agent/conftest/runner"
 )
